# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text or GIFs and
    easily stay in touch through your private chats or group discussion.
  version: "0.1"
servers:
  - url: http://localhost:3000
...
paths:
...
  /users/{userId}/username:
    put:
      tags:
        - User
      summary: change username
      description: |
        change your Username to a new, unused one
      operationId: setMyUserName
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        description: "send the request with the new username"
        required: true
        content:
          application/json:
            schema:
              description: "object with new username as a string"
              type: object
              properties:
                username:
                  $ref: "#/components/schemas/Name"
              required:
                - username
      responses:
        "202":
          description: "username changed successfully"
          content:
            application/json:
              schema:
                description: new Username
                type: object
                properties:
                  username:
                    $ref: '#/components/schemas/Name'
                required:
                  - username
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          description: "Conflict: The new username is already taken."
          content:
            text/plain:
              schema:
                description: username already taken
                type: string
                example: "username already taken"

        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/photoProfile:
    put:
      tags:
        - User
      summary: change photo profile
      description: send the request with the new photo profile
      operationId: setMyPhoto
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        description: a photo to be set
        required: true
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/Image"
      responses:
        "201":
          description: Profile photo updated successfully.
          content:
            application/json:
              schema:
                description: Unique image code
                type: object
                properties:
                  imageCode:
                    type: string
                    minLength: 27
                    maxLength: 27
                    pattern: "^[0-9a-zA-Z]*?"
                    example: "1ThDPX91uDwB1yxw5jYRxUIVs1W"
                required:
                  - imageCode
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/conversations:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - Chats
      summary: Conversations list
      description: Get the list of conversations of a specific user
      operationId: getMyConversations
      responses:
        "200":
          description: List of conversations retured successfully
          content:
            application/json:
              schema:
                description: array of conversation of a user and total of conversation
                type: object
                properties:
                  conversations:
                    description: array of conversation with all associated informations
                    type: array
                    items:
                      $ref: "#/components/schemas/ConversationInfo"
                    minItems: 0
                    maxItems: 250
                required:
                  - conversations
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Chats
      summary: start a new conversation
      description: |
        used to create a new group or add new user in your conversation list
      operationId: startConversation
      requestBody:
        description: |
          we need to know if you are creating a group or adding a new user to your conversation list.
          if you are creating a group then we will need the name but if you are adding a person then we will need then person's name.
        required: true
        content:
          application/json:
            schema:
              description: need a name and a type of conversation
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
                typeConversation:
                  type: string
                  enum: [private,group]
              required:
                - name
                - typeConversation
      responses:
        "201":
          description: "conversation created successfully"
          content:
            application/json:
              schema:
                description: "response have 3 field message from server, name and type convesation"
                type: object
                properties:
                  identifier:
                    $ref: "#/components/schemas/Identifier"
                required:
                  - identifier
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "400":
          $ref: '#/components/responses/BadRequest'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/groupName:
    put:
      tags:
        - Groups
      summary: change group name
      description: set a new name for a group, must be a member group
      operationId: setGroupName
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: a new group name
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
              required:
                - name
      responses:
        "200":
          description: group name changed successfully
          content:
            application/json:
              schema:
                properties:
                  groupName:
                    $ref: "#/components/schemas/Name"
                required:
                  - groupName
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/groupPhoto:
    put:
      tags:
        - Groups
      summary: change group photo
      description: set a new image as photo profile for a group, you must be in the group
      operationId: setGroupPhoto
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        required: true
        content:
          image/png:
            schema:
              $ref: "#/components/schemas/Image"
          image/jpeg:
            schema:
              $ref: "#/components/schemas/Image"
      responses:
        "201":
          description: Profile photo updated successfully.
          content:
            application/json:
              schema:
                description: Unique image code
                type: object
                properties:
                  imageCode:
                    type: string
                    minLength: 27
                    maxLength: 27
                    pattern: "^[0-9a-zA-Z]*?"
                    example: "1ThDPX91uDwB1yxw5jYRxUIVs1W"
                required:
                  - imageCode
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    conversationId:
      description: identifier for a single conversation
      name: conversationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    userId:
      description: identifier for a single user
      name: userId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
  schemas:
    Name:
      title: UserName/GroupName
      description: schema used to rappresent username or group name
      type: string
      pattern: "^[a-zA-Z0-9]*$"
      minLength: 3
      maxLength: 25
      example: "Deadpool69"
    Identifier:
      title: ID
      description: "A unique identifier for the user, in the specified format."
      type: string
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      minLength: 15
      maxLength: 15
      example: "123e4567-e89b-12d3-a456-426614174000"
    Image:
      title: Image
      description: Image binary format as a string
      type: string
      format: binary
      minLength: 1
      maxLength: 20971520 # 20MB
...
    ConversationInfo:
      description: |
        list of conversation with other users or with groups, sorted in reverse chronological order.
        Each element in the list must display the username of the other person or the group name, the user profile photo or the group photo, the date and time of the latest message, the preview (snippet) of the text message, or an icon for a photo message.
      type: object
      properties:
        conversationId:
          $ref: "#/components/schemas/Identifier"
        conversationType:
          description: type of conversation
          type: string
          enum: [private,group]
        conversationName:
          $ref: "#/components/schemas/Name"
        photoProfileCode:
          description: "photo code (unique)"
          type: string
          minLength: 27
          maxLength: 27
          pattern: "^[0-9a-zA-Z]*?"
          example: "1ThDPX91uDwB1yxw5jYRxUIVs1W"
        lastMessageTimeStamp:
          description: timestamp of the last message of conversation
          type: string
          format: date-time
          example: "2024-12-23T09:00:00Z"
          pattern: '^\d{4}-\d{2}-\d{2}T-\d{2}:\d{2}:\d{2}Z$'
          minLength: 20
          maxLength: 20
        lastMessagePreview:
          description: snippet of the last message in coversation
          type: string
          maxLength: 50
          example: "hey buddy, How ..."
      required:
        - conversationId
        - conversationType
        - conversationName
        - photoProfileId
        - lastMessageTimeStamp
        - lastMessagePreview
...
  responses:
...
    BadRequest:
      description: syntax of the request was not compliant with docs, check parameters
      content:
        text/plain:
          schema:
            description: bad request string message
            type: string
            example: Bad Request
    #401
    Unauthorized:
      description: authorization missing or not valid
      content:
        text/plain:
          schema:
            description: Unauthorized string message
            type: string
            example: Unauthorized
    #403
    Forbidden:
      description: server will not accept request due to lacking of client permissions
      content:
        text/plain:
          schema:
            description: Forbidden string message
            type: string
            example: Forbidden
    #404
    NotFound:
      description: resource not found
      content:
        text/plain:
          schema:
            description: Not Found string message
            type: string
            example: Not found
    #409
...
    InternalServerError:
      description: the server encountered an error while processing the request
      content:
        text/plain:
          schema:
            description: Internal Server Error string message
            type: string
            example: Internal Server Error
  securitySchemes:
    TokenAuth:
      description: |
        User authentication with the token generated by backend application
      type: http
      scheme: bearer

# Apply security scheme globally, disabling it explicitly when unnecessary with security: [].
security:
  - TokenAuth: []